refactor: consolidate BlogPost interface into shared type
