+++ types/blog.ts
@@ -1,0 +1,18 @@
+export interface BlogPost {
+  slug: string
+  title: string
+  author: string
+  date: string
+  image: string
+  excerpt: string
+  content: string
+  featured: boolean
+}
+
+export interface PaginatedPosts {
+  posts: BlogPost[]
+  totalPages: number
+  currentPage: number
+  hasNextPage: boolean
+  hasPrevPage: boolean
+}
